requirements.txt
streamlit==1.37.1
google-generativeai==0.7.2
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
